ai-agent/requirements.txt
streamlit
python-dotenv
requests
beautifulsoup4
duckduckgo_search
wikipedia
